Add Report for generating success and failure reports.
